+++ ChangeLog.md
@@ -11,6 +11,7 @@ Eventum Issue Tracking System
 - Add html_charset to be APP_CHARSET (Elan Ruusamäe, LP#741768)
 - Use password_hash family functions for password hashing (Elan Ruusamäe, GH#77)
 - Unify size of pri_id in databases (Elan Ruusamäe, LP#1450152)
+- Handle mbstring function overload (Elan Ruusamäe, LP#1494732)
 
 2015-10-13, Version 3.0.3
 -------------------------